+++ R/fastcca.R
@@ -59,11 +59,16 @@ setMethod("show", "fastcca", function(object) {
 #' pop <- LifeCycleSavings[, 2:3]
 #' oec <- LifeCycleSavings[, -(2:3)]
 #'
-#' fastcca(pop, oec)
-#'
+#' # fit CCA
+#' fit <- fastcca(pop, oec)
+#' fit
+#' 
+#' # predict second dataset given first
+#' y.pred <- predict(fit, X = pop)
+#
 #' @importFrom Rfast cora colsums
 #' @export
-fastcca = function(X, Y, k = NULL, k.x=min(dim(X)), k.y=min(dim(Y)), lambda.x=NULL, lambda.y=NULL, svd.method = c("svd", "irlba", "pcaone") ){
+fastcca <- function(X, Y, k = NULL, k.x=min(dim(X)), k.y=min(dim(Y)), lambda.x=NULL, lambda.y=NULL, svd.method = c("svd", "irlba", "pcaone") ){
 
   # checks
   stopifnot("k must be positive" = k > 0)
@@ -144,7 +149,7 @@ fastcca = function(X, Y, k = NULL, k.x=min(dim(X)), k.y=min(dim(Y)), lambda.x=NU
 
   rho.mod <- dcmp$d
 
-  res = list( 
+  res <- list( 
         dims = c(n=nrow(X), p1 = ncol(X), p2 = ncol(Y)),
         n.comp = k,
         rho.mod = rho.mod,
@@ -155,7 +160,124 @@ fastcca = function(X, Y, k = NULL, k.x=min(dim(X)), k.y=min(dim(Y)), lambda.x=NU
         x.coefs = x.coefs, 
         y.coefs = y.coefs, 
         x.vars = x.vars, 
-        y.vars = y.vars)
+        y.vars = y.vars,
+        x.mu = attr(X_scaled, "mu"),
+        y.mu = attr(Y_scaled, "mu"),
+        x.sd = attr(X_scaled, "sd"),
+        y.sd = attr(Y_scaled, "sd"))
 
   new("fastcca", res)
 }
+
+#' Predict one dataset from another
+#' 
+#' Using CCA fit, predict one dataset from another
+#' 
+#' @param object model fit from \code{fastcca()}
+#' @param X first dataset 
+#' @param Y second dataset 
+#' @param ... other arguments, not used
+#'
+#' @details Specify which dataset to map from using \code{X} and \code{Y}.  
+#'
+#' @examples
+#' pop <- LifeCycleSavings[, 2:3]
+#' oec <- LifeCycleSavings[, -(2:3)]
+#'
+#' # fit CCA
+#' fit <- fastcca(pop, oec)
+#' 
+#' # predict Y given X
+#' y.pred <- predict(fit, X = pop)
+#' head(y.pred)
+#' 
+#' # predict X given Y
+#' x.pred <- predict(fit, Y = oec)
+#' head(x.pred)
+#
+#' @export
+setMethod("predict", "fastcca", function(object, X, Y,...) {
+
+  if( !missing(X) & !missing(Y) ){
+    stop("Only one of X and Y can be specified")
+  }
+
+  if( !missing(X) ){
+    res <- predict_from_X(object, X)
+  }
+
+  if( !missing(Y) ){
+    res <- predict_from_Y(object, Y)
+  }
+
+  res
+})
+
+
+
+#' @importFrom MASS ginv
+#' @importFrom Rfast eachrow
+predict_from_X <- function( fit, X ){
+
+  # check that column names of X are the same as in 
+  # the training data
+  if( ! identical(rownames(fit$x.coefs), colnames(X)) ){
+    stop("column names of X doesn't the training data ")
+  }
+
+  X <- as.matrix(X)
+
+  # scale input data
+  X_scaled <- eachrow(X, fit$x.mu, "-")
+  X_scaled <- eachrow(X_scaled, fit$x.sd, "/")
+
+  # project X onto latent space
+  # project from latent space to Y
+  y.pred <- X_scaled %*% (fit$x.coefs %*% ginv(fit$y.coefs))
+  
+  # rescale to Y
+  y.pred <- eachrow(y.pred, fit$y.sd, "*")
+  y.pred <- eachrow(y.pred, fit$y.mu, "+")
+
+  colnames(y.pred) <- rownames(fit$y.coefs)
+  rownames(y.pred) <- rownames(X)
+
+  y.pred
+}
+
+
+#' @importFrom MASS ginv
+#' @importFrom Rfast eachrow
+predict_from_Y <- function( fit, Y ){
+
+  # check that column names of Y are the same as in 
+  # the training data
+  if( ! identical(rownames(fit$y.coefs), colnames(Y)) ){
+    stop("column names of Y doesn't the training data ")
+  }
+
+  Y <- as.matrix(Y)
+
+  # scale input data
+  Y_scaled <- eachrow(Y, fit$y.mu, "-")
+  Y_scaled <- eachrow(Y_scaled, fit$y.sd, "/")
+
+  # project Y onto latent space
+  # project from latent space to X
+  x.pred <- Y_scaled %*% (fit$y.coefs %*% ginv(fit$x.coefs))
+  
+  # rescale to Y
+  x.pred <- eachrow(x.pred, fit$x.sd, "*")
+  x.pred <- eachrow(x.pred, fit$x.mu, "+")
+
+  colnames(x.pred) <- rownames(fit$x.coefs)
+  rownames(x.pred) <- rownames(Y)
+
+  x.pred
+}
+
+
+
+
+
+